split up scatter methods into separate files
